refactor(journey): moved step metadata related work into the journey.store file
- Moved metadata function calls from the stage components into the journey.store file
- Improved types
