+++ src/lib/journey/callbacks/kba/kba-create.svelte
@@ -22,7 +22,7 @@
   export const stepMetadata: Maybe<StepMetadata> = null;
 
   export let callback: never;
-  export let callbackMetadata: CallbackMetadata;
+  export let callbackMetadata: Maybe<CallbackMetadata>;
   export let style: Style = {};
 
   const Input = style.labels === 'stacked' ? Stacked : Floating;
@@ -116,7 +116,7 @@
 
   $: {
     inputArr = typedCallback?.payload?.input;
-    inputName = typedCallback?.payload?.input?.[0].name || `kba-${callbackMetadata.idx}`;
+    inputName = typedCallback?.payload?.input?.[0].name || `kba-${callbackMetadata?.idx}`;
     inputNameQuestion = inputName;
     inputNameAnswer = Array.isArray(inputArr) && inputArr[1].name;
     prompt = typedCallback.getPrompt();
@@ -179,7 +179,7 @@
   {#if displayCustomQuestionInput}
     <Input
       isFirstInvalidInput={false}
-      key={`kba-custom-question-${callbackMetadata.idx}`}
+      key={`kba-custom-question-${callbackMetadata?.idx}`}
       label={interpolate('customSecurityQuestion')}
       showMessage={false}
       message={interpolate('inputRequiredError')}
@@ -189,8 +189,8 @@
   {/if}
 
   <Input
-    isFirstInvalidInput={callbackMetadata.isFirstInvalidInput}
-    key={inputNameAnswer || `kba-answer-${callbackMetadata.idx}`}
+    isFirstInvalidInput={callbackMetadata?.isFirstInvalidInput || false}
+    key={inputNameAnswer || `kba-answer-${callbackMetadata?.idx}`}
     label={interpolate('securityAnswer')}
     showMessage={false}
     message={interpolate('inputRequiredError')}
