+++ src/lib/journey/callbacks/polling-wait/polling-wait.story.svelte
@@ -9,11 +9,12 @@
     SelfSubmitFunction,
     StepMetadata,
   } from '$journey/journey.interfaces';
+  import type { Maybe } from '$lib/interfaces';
 
   export let callback: never;
-  export let callbackMetadata: CallbackMetadata;
+  export let callbackMetadata: Maybe<CallbackMetadata>;
   export let selfSubmitFunction: SelfSubmitFunction;
-  export let stepMetadata: StepMetadata;
+  export let stepMetadata: Maybe<StepMetadata>;
 </script>
 
 <Centered>
