+++ src/lib/journey/journey.store.ts
@@ -7,8 +7,10 @@ import type { JourneyStore, JourneyStoreValue, StackStore, StepTypes } from './j
 import { interpolate } from '$lib/_utilities/i18n.utilities';
 import {
   authIdTimeoutErrorCode,
+  initCheckValidation,
   shouldPopulateWithPreviousCallbacks,
-} from './_utilities/step.utilities';
+} from './stages/_utilities/step.utilities';
+import { buildCallbackMetadata, buildStepMetadata } from '$journey/_utilities/metadata.utilities';
 
 function initializeStack(initOptions?: StepOptions) {
   const initialValue = initOptions ? [initOptions] : [];
@@ -61,6 +63,7 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
     completed: false,
     error: null,
     loading: false,
+    metadata: null,
     step: null,
     successful: false,
     response: null,
@@ -95,6 +98,7 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
       completed: false,
       error: null,
       loading: true,
+      metadata: null,
       step: prevStep,
       successful: false,
       response: null,
@@ -135,6 +139,8 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
       /**
        * SUCCESSFUL CONTINUATION BLOCK
        */
+      const callbackMetadata = buildCallbackMetadata(nextStep, initCheckValidation());
+      const stepMetadata = buildStepMetadata(callbackMetadata);
 
       // Iterate on a successful progression
       stepNumber = stepNumber + 1;
@@ -143,6 +149,10 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
         completed: false,
         error: null,
         loading: false,
+        metadata: {
+          callbacks: callbackMetadata,
+          step: stepMetadata,
+        },
         step: nextStep,
         successful: false,
         response: null,
@@ -157,6 +167,7 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
         completed: true,
         error: null,
         loading: false,
+        metadata: null,
         step: null,
         successful: true,
         response: nextStep.payload,
@@ -235,6 +246,9 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
        * the final result to the user.
        */
       if (restartedStep.type === StepType.Step) {
+        const callbackMetadata = buildCallbackMetadata(restartedStep, initCheckValidation());
+        const stepMetadata = buildStepMetadata(callbackMetadata);
+
         set({
           completed: false,
           error: {
@@ -244,6 +258,10 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
             step: prevStep?.payload,
           },
           loading: false,
+          metadata: {
+            callbacks: callbackMetadata,
+            step: stepMetadata,
+          },
           step: restartedStep,
           successful: false,
           response: null,
@@ -253,6 +271,7 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
           completed: true,
           error: null,
           loading: false,
+          metadata: null,
           step: null,
           successful: true,
           response: restartedStep.payload,
@@ -267,6 +286,7 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
             step: prevStep?.payload,
           },
           loading: false,
+          metadata: null,
           step: null,
           successful: false,
           response: restartedStep.payload,
@@ -302,6 +322,7 @@ export function initialize(initOptions?: StepOptions): JourneyStore {
       completed: false,
       error: null,
       loading: false,
+      metadata: null,
       step: null,
       successful: false,
       response: null,
